% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctt_eval_function.R
\name{ctt_eval}
\alias{ctt_eval}
\title{Title of Function: ctt_eval}
\usage{
ctt_eval(data)
}
\arguments{
\item{x}{Must be a data frame or matrix.}
}
\value{
Will return two tables of statistics to evaluate your scale: 1) item-level and 2) scale-level
}
\description{
Function applies Classical Test Theory statistics for scale and item evaluation.
}
